(PUP-7736) Conditionally require 'puppet'
The generate_ast_model rake task requires puppet, which requires facter.
At build-time, we don't always have facter gems available (but we also
don't need the tasks defined in generate_ast_model). To make it possible
to build puppet, we need to not load the generate_ast_model tasks if
facter is not available.