(PUP-5506) making the source parameter authoritative for the gem package provider
Prior to this commit, the 'install_options' attribute was passed to the gem
command after the 'source' attribute, preventing the use of '--clear-sources'
in 'install_options', as '--clear-sources' is positionally dependent.

With this commit, the 'install_options' attribute is passed first.

This was identified and resolved in MODULES-4815.

While the subject of the ticket is 'making the source parameter authoritative',
in MODULES-4815 we decided to maintain backwards compatibility and not imply
'--clear-sources' when the 'source' attribute is specified.