(PUP-1391) puppet device cannot create certs when run as root
Prior to this commit, puppet device could not write to the privatekeydir
and certdir for devices.

drwxr-x---.  3 root      root        18 Jun 28 22:55 devices

With this commit, the $vardir/devices directory’s owner and and group are set
to the service user, so that the service user can traverse the directory.

drwxr-x---.  3 pe-puppet pe-puppet   18 Jun 28 22:55 devices