(PUP-7738) Ensure that Puppet can be used with semantic_puppet gem.
This commit adds an arity check to be able to call the method
`SemanticPuppet::VersionRange#parse` with one or two arguments from the
`Puppet::Module#parse_range` method.

When the arity is found to be just 1, a check is made if the gem version
of the semantic_puppet gem is < 1.0.0, in which case it is assumed that
the parse method is incapable of creating a range that adheres to strict
semantics. A warning is logged if a request for a strict range is made.
Conversely, if the gem version is >=1.0.0, it is assumed that strict
semantics will be honored and a warning is issued if a request for a
non-strict range is made.