(PUP-7732) Audit Acceptance Tests - PluginSync tests
This commit adds audit tags to tests in order to facilitate the
proper application of test tiering and layering in the future.
The `audit` tags are intended to be removed once the corresponding
work has been completed.

Also, the `server` tag has been added to any tests that should be
executed in the context of `puppetserver` and not in the context
of `puppet` or `puppet-agent` testing.