(PUP-7738) Ensure that option :module_repository is propagated in Search
It appears that the while the unit test for the `search` face that
tested if the option `:module_repository` was accepted, it did not check
if it was also propagated to the `Puppet::Forge` instance. It wasn't,
and as a consequence, the test failed when the default repository was
propagated in an attempt to simulate the default argument assignment
when passing no arguments to `Puppet::Forge#new`.

This commit therefore fixes two things:
1. It ensures that the `:module_repository` option is propagated when
   the `Puppet::Forge` is instantiated.
2. It updates the test to check that the correct repository is
   propagated when the instance is created.