(PUP-7753) Ensure that timestamps using space separator can be parsed.
This commit adds default formats to enable parsing of timestamps where
the date and time part are separated by space. It also adds tests to
assert that such timestamps are fully parsed.

An attempt was made to check if a string is fully parsed so that partly
parsed strings would raise an error. Unfortunately, the parser in
`DateTime#strptime` that Puppet currently relies on, does not make this
possible. It parses up to the point where a format is fully satisfied
and then returns the result without detecting trailing garbage.

A test to detect such garbage is added in this commit, but since it
would require a complete rewrite of the parser, it's currently set
to "pending".