(PUP-7757) Raise errors unless Timespan#parse consumes all input
This commit changes the Timestamp#parse method so that it uses the
method `DateTime#_strptime` instead of `DateTime#strptime` internally.
The `_strptime` returns a `Hash` instead of a `DateTime` and this hash
will contain the undocumented field `leftover` in case the parsed string
was not consumed in full. In Puppet, it's desirable to treat such data
as an error since it means that only part of a string represented the
desired `Timestamp`.

Some extra logic was added to process the parse result since it's now
in the form of a `Hash` instead of a `DateTime`. Also, the previous
begin/rescue was no longer needed.since the `_strptime` returns `nil`
instead of raising an ArgumentError for unparsable input.