(PUP-7770) Show all installed packages in pip package provider
By default pip hides "development" dependencies in the output of
`pip freeze`.  In pip 8.0.3 the `--all` flag was added to pip to force
those packages to be shown.  This adds that flag to the pip command if
pip is new enough.  In addition this makes the hack to add pip to the
installed package list unnecessary.