(PUP-4755) prefer explicit [main] section header
The `puppet config set` command previously allowed and preferred
that the [main] section header be omitted, and [main] settings be
written in the ini file as global.

This commit rejects that standard, preferring instead that the [main]
section header always be explicitly written in the file when `puppet
config` is used.