(PUP-7786) Do not perform English string checks when running in a Japanese locale
This commit prevents certain string-match assertions from being run when
testing on a Japanese host, since those checks will fail when there are
translations in place. This uses a Beaker host option to detect when we
are running with Japanese, since currently Beaker cannot tell the
difference between our English and Japanese Windows templates.

In order for these confines to take affect, use a test target string
suffixed with `{locale=ja}`. This is not intended to be a permanent
solution, merely a stop-gap measure to allow testing to continue as
Japanese translations are landing.