(PUP-5047) POSIX Exec throws error when group has no name
On Macs which are domain-joined, lookups between group ID and name can
fail if the domain isn't accessible at that moment. In these situations,
lookups on files and other objects return a GID for the user that is
already an integer, which is then passed to convert_xid, which tries to
re-lookup that GID and fails, returning an "Invalid group: ###" error.

This PR bails early if the user/group ID is already numeric, and thus
doesn't need to be looked up.