(PUP-7802) Make hiera_xxx functions unaffected by data_binding_terminus
Prior to this commit, all global lookups would be affected by the
`data_binding_terminus` setting. This commit ensures that this is no
longer the case when the lookup origins from a `hiera`, `hiera_array`,
`hiera_hash`, or `hiera_include` function.

The change is made to ensure backward compatibility with Hiera 3 where
the `hiera_xxx` calls would go directly to the classic Hiera and thus
not be affected by the `data_binding_terminus`.