(PUP-7804) Do not use server setting when failover fails
Previously, when failover was attempted but unsuccessful, the
agent would default to using the `server` setting. This is
incorrect, since that setting might be at its default value,
or be set to the same value as one of the server_list entries.

Instead, we should not attempt to connect to any server if failover
was not successful. This allows the agent to "fail fast", instead of
timing out trying to connect to an incorrect server.