(PUP-6241) Only expect incorrect File exception on JRuby 1.7
Previously, our tests for an environment bug on JRuby 1.7 would execute
on any Puppet Server instance.

Since we have a regular JRuby 9k version of our Puppet Server jobs this
means this test, which expects a bug that only exists in JRuby 1.7
fails.

This patch updates the acceptance test to skip if Puppet Server is using
JRuby 9k.