(maint) Restore attempting to remove files even if backup fails
The fix for PUP-6966 introduced a new error, halting file removal if
backup could not be performed. While ultimately this seems like the
correct behavior, we will need to further investigate the implications
of this logic change. For the time being, this commit changes the
error to a warning, allowing the code to proceed to try to remove the
file even if backup has failed.