(PUP-7807) Add mount options on AIX remount
Prior to this commit, the mount resource on AIX would not remount NFS
shares with the mount options specified. Since AIX did not read the NFS
options from the `/etc/filesystems`, the share would be remounted
without options.

This commit adds a simple check to ensure all AIX remounts are done with
the specified options including JFS2.