(PUP-7818) Prefetch AIX packages when not all have sources
Previously, the prefetch for the `aix` package provider assumed
that all resources would have a `source` parameter. While this
parameter is required in order to *install* a package, it is
not needed to remove one, so it's possible some package resources
might lack a value.

This fixes the prefetch to skip any empty source values, allowing
prefetch to succeed when some packages do not have a source.