(PUP-7824) Fix bad problem with unanchored end in string format regexp
The regexp used for checking format specifiers for the string formatter
was not anchored at the end. Also, the regexp was duplicated. It was
specified once in string_formatter.rb and then again in as a parameter
type in types.rb.

This commit adds the anchor and removes the duplification.