(PUP-7829) Ensure that Variant parameters are types
The `TypeParser` did not check what the parameters used when creating
a `Variant` type were types. Consequently, some variants created
internally in dispatchers used for the function `new` on `Timespan`
and `SemVerRange`, were incorrect. The faulty declarations were also
copied into the corresponding documentation of the `new` method.

Incidentally, the code did work anyway since the type, when created
from the `TypeParser`, will do a `#to_type` on the argument, thus
converting the keyword `default`, or an integer value into it's type.