(PUP-7829) Fix error when formatting errors for TypeAliases with variants
The TypeFormatter would incorrectly swap the expected type in
a description for the original alias although the message description
was for something nested contained in the resolved type of the alias.

This commit ensures that the replacement only happends when the sizes
of the path to the error description are equal.