(PUP-7768) Enable use of heredoc in a type description
Prior to this commit, the `TypeParser` would raise an exception when it
encountered a `HeredocExpression`. As a consequence, no strings could
be entered as heredoc in a type description.

This commit adds the methods necessary to make the `TypeParser`
understand literal heredoc. Interpolations are not allowed (goes for
both heredoc and double quoted strings).