(PUP-7834) Use custom implementation instead of YAML.safe_load
The `YAML.safe_load` was introduced in Ruby 2.1 and can therefore not
be directly used in Puppet. This commit changes all calls to that method
to instead use `Puppet::Util::Yaml.safe_load`. That method in turn,
will either call `YAML.safe_load` or use a custom implementation of that
method depending on the current version of the ruby runtime.