(PUP-7840) Add task_file method to modules.
Previously, modules had no way to take the name of a task file and
return the full path if it existed; this required iterating through the
modules' tasks and all the files of each task to find a match, which is
inefficient. For the purposes of file serving, it's better to construct
the path where the file would exist and check the filesystem directly.
Add a `task_file` instance method to modules that takes this approach.