(PUP-7844) Return -1 for exit_status for interrupted runs
If a puppet run is interrupted, the local puppet report in
last_run_report will be incomplete, which will mean that the metrics
instance variable will portentially be the empty hash. Because the
lookups in `exit_status` against `metrics` is not safe, it could raise
an exception. This commit updates the `exit_status` method to check to
make sure `metrics` has the expected keys and if not to return -1
instead of raising an exception.