(PUP-7845) Implement leaf certificate revocation checking
Puppet has always performed chain revocation checking when revocation
checking was enabled, but has never been able to load more than a single
CRL. Because of this Puppet could only be used with a root CA if
revocation checking was desired, and if intermediate CAs were in used
revocation had to be entirely disabled.

This commit sidesteps the issue of loading multiple CRLs by allowing
certificate revocation checking to be performed on the leaf certificate
only. This allows users to revoke certificates with an intermediate CA
that Puppet is managing, which is a definite improvement over not
allowing any revocation checking at all.

For the sake of compatibility certificate_revocation still defaults to
full chain checking; users must opt in to leaf revocation checking.