(PUP-7813) Ensure yum plugins don't break yum
The yum plugin API allows for plugins to send output via info or
error messages to stdout. This can break assumptions the yum provider
is making in terms of the output from commands like check-update.
This adds a check to ensure that when parse_updates is called, the
second argument in the tuple matches an RPM version and skips over it
if not. This should help guard against multiple types of bad output.