(PUP-7852) Add attribute 'constants' to the Object type description
This commit adds a new member `constants` the `Object` type. The
`constants` member is similar to the `attributes` member in that its
value is a hash of attributes. The specification of each attribute under
`constants` is however only a value. It's type is inferred from this
value and the kind is always 'constant'.

Overrides are automatically allowed if the parent attribute is also a
constant with a type that is assignable from the overriding type.