(PUP-7855) Ensure proper Data conversion the special default value
Before this commit, the ToDataConversion would serialize the special
`default` value using rich data semantics even when `rich_data` was
set to false. This commit ensures that a) this only happens when rich
data is enabled, and b) when it happens, a proper warning is logged.

A new issue was added to avoid logging info about the symbol `:default`
and instead mention the "special value default" (the use of the symbol
is an an internal concern of the ruby implementation).