(PUP-7858) Isolate and externalize all `puppet help` summary texts
Some summary texts for subcommands appearing in `puppet help` missed
being translated. For some of these, that was because the string was
embedded in a larger help text (which was not intended for translation)
and extracted on the fly. This commit isolates all of these summary
texts and makes them available via a `summary` method. This allows them
to be marked for translation.

The summary method return an empty string by default. If users of puppet
have created their own applications without summaries, the `help`
command will still attempt to extract the summary text from the help
body.