(PUP-7862) Ensure variant assignability to Data and RichData
Before this commit, only subclasses to `PScalarDataType` were considered
assignable the `ScalarData` type. This commit ensures that all variants
containing assignable types (i.e. `Optional`, `NotUndef`, or `Variant`)
are assignable (with the exception of `NotUndef[Undef]`, which isn't
assignable to anything).

The `Scalar` datatype which experienced the same problem, is also fixed
by this commit.