(PUP-6364) Fix `ensure => latest` on AIX
Previously, the latest version of a package in AIX was not
being returned when `ensure => latest` was used. In the calculation
of the latest version of the package, `latest_info` was being assigned
to `nil` rather than the package itself. This resulted in whatever came
first in the TOC file being the package that was installed/used. In the
case where one package would be installed and an update was released,
the first entry in the .toc file would be the old package, so the update
would not occur.

This simply fixes the logic to calculate the package by version, rather
than falling back to first in the .toc file.