(PUP-7875) Add a mocking version of run_task
This adds a mocking version of run_task where it is possible to simulate
running of a task by giving it an Object instance that has the
attributes simulated_exit_code and simulated_result. All other
attributes are ignored.

The function logs information at notice level.

This behavior will be replaced when there are real Task instances.