(PUP-7864) Make it possible to pass loader to Object.new
This change is to accommodate the very special occasion when an `Object`
is created from an already resolved hash and hence never will receive
a call to it's resolve method. When that situation occurs, the created
type will never get its loader assigned.

This change gives the caller a chance to pass the loader directly in the
call to Object.new.