(PUP-7864) Change the root path loaders to not point into 'lib'
Prior to this commit, the environment and module loaders all had a path
that pointed into the <root>/lib directory. While this might have made
sense in the past (when everything loaded actually resided under lib),
it doesn't today, when only the ruby loader uses the 'lib' folder.

This commit ensures that the loaders instead point directly to the root
and that all `SmartPath` subclasses take this into account.