(PUP-7864) Add the Task type to the Puppet Type System
Adds the Object subtypes Task and GenericTask to the Puppet Type System.
The Task is a ruby implementation with methods to produce the task JSON
and the path of the exectutable. It also has constants like
'supports_noop' and 'input_method' needed by the task runner.

A `GenericTask` is like `Task` but instead of presenting all attributes
in a hash, it has one single attribute that in itself represents the
hash.