(PUP-7864) Make it possible for a loader to find several files
A loader normally finds one single file that represents the thing to be
loaded and then passes that to an instantiator. This functionality must
be improved so that multiple files can be found and passed instead.

This commit adds the functionality needed to create a `SmartPath` that
can find multiple files and then pass them to the instantator. In this
scenario, it's up to the instantiator to read the files.