(PUP-7864) Add Puppet Type Loader for Tasks
This commit adds a loader that finds files under the 'tasks' folder
in a module root directory. The loader will expect to find exactly one
file that matches the expected task name and has an optional or
arbitrary file extension that is different from '.json'. That file may
optionally be accompanied by file that has the same name and a '.json'
extension. The json latter is then the metadata that describes the task
type.