(PUP-7864) Fix glitch affecting nested subclassing of Object types
The RubyGenerator assumed that a generated class could leave out class
methods in case those methods were exactly equal to the same methods of
the class generated for the super type. This worked fine as long as the
type of an implementation class inherited directly from the Object type
but it didn't work when the actual implementation was provided in Ruby
and inherited from another implementation, also provided in Ruby (as
when GenericTask (which is an Object type) inherits from Task (which
also is an Object type).

This commit changes the generator so that it always generates the class
methods.