(PUP-7864) Remove incorrect exclusion of _pcore_init_hash method
The `RubyGenerator` would generate an incorrect `#_pcore_init_hash`
for `Object` implementations that didn't have any attributes. This
commit ensures that no such method is generated (rely on super) and
that a super that includes `PuppetObject` but doesn't override, returns
a modifiable hash (specializations often add things to this hash before
it is returned).