(PUP-7879) Add a "puppet script" application that is similar to apply
This adds the application "puppet script" that is similar to
"puppet apply", but it does not support building a catalog, nor does it
support using classes or resources (other than the main (a.k.a '')
class) that represents "top scope logic".

It does this by using a "ScriptCompiler" directly rather than going via
an indirection.

The Script compiler does not set individual facts as top scope variables
and it does not create the class "settings", but it makes all settings
available as in regular puppet.

All facts are for the local node where "puppet script" is running - they
are available in $facts and $trusted.