(PUP-7885) Ensure that Enum data type only accepts string parameters
A parameterized `Enum` must only accept parameters of type string
(or identifier, since it's converted to a string) but prior to this
commit it was possible to use arbitrary parameters such as integers if
the string was parsed/evaluated by the internal TypeParser. This commit
adds the check needed to prevent this from happening.