(PUP-7911) Avoid gem install conflict of this branch and other puppet installs
We need to make sure that when this branch gets installed via a Ruby gem,
that we do not interfere with any other Puppet installations that the user
might already have. We do this by not mentioning 'lib/' in the require_path
of our gemspec. This effectively hides everything in 'lib/' from gem and
will keep ruby from ever satisfying a require of 'puppet/<somewhere>' with
files from this gem.

Code that consciously wants to load this gem, needs to instead require
'puppet_bolt'. That manipulates $LOAD_PATH so that 'lib/' does become
available. After that, any 'require "puppet/<somewhere>"' will resolve to a
file from this gem.

Note that there are two ways to build a Puppet gem: 'gem build' (driven by
.gemspec) will continue to produce a puppet gem; that gem is used by Travis
and Appveyor, and we do not change it to make sure those tests still
run. For release, we use 'rake package:gem' which is driven off
ext/project_data.yaml. We only modify that file.