(PUP-7890) Add discovery to the Puppet Loader system
This commit adds a `#discover` method to all pcore loaders. This new
method can be used to discover all entities that the loader has access
to. The type of entity to search for must be provided. A filter can
also be provided in the form of a block.

Current limitations:
- All entities that are discovered will be loaded. This has pros and
  cons. The pro is that once loaded, the type will be cached in memory.
  The con is that loading and keeping all entities in memory might be
  resource demanding.
- Resource types are currently not loaded.

A solution for both limitations would be to provide a utility program
that would execute separate from the puppet runtime to maintain an
index of everything that can be loaded.