(PUP-7879) Refactor ScriptCompiler to remove dependency on node
This refactors the ScriptCompiler so that it is free from any node
concerns. The ScriptCompiler simply compiles after having been given an
environment (and a node_name to use when reporting errors).

This makes it more flexible as support for setting things like $trusted,
$facts, $settings, etc. depends on the use case. Initialization of those
are now moved to the Script application.

Additional tweaks and cleanup is also included.