(PUP-7902) Ensure that empty heredoc lines are not trimmed
Before this commit, the regexp that was used to trim off the last
newline (or cr-nl) was not a multi-line regexp and used the '$' (end of
line) which resulted in that all empty lines where removed. This commit
changes the regexp into a multiline that uses \z (end of string), thus
ensuring that only the last occurrence is removed.