(PUP-7905) Make it possible to override polymorphic Checker4_0
The `Checker4_0` class cached it's polymorphic "check" visitor in a
class variable. This doesn't work in case the class is overridden and
there is a chance that instances of both the base class and the
overridden class can co-exist in the same ruby runtime.

This commit moves the cached visitor to a class instance variable to
ensure that differnt classes have different visitors.