(PUP-7908) Remove use of 'main' class for ScriptCompiler
This commit changes the `ScriptCompiler#evaluate_main` so that it no
longer creates a class with an empty name (a.k.a. 'main').

The method `Environment#perform_initial_import` was moved to the
`Loaders` class and rewritten to no longer create wrappers for the
version 3 AST. Instead, it directly returns an array of Program
instanses.

The `ScriptCompiler#evaluate_main` calls the new method on the `Loaders`
instance to obtain the programs to evaluate. It then evaluates them in
sequence and returns the result of the last evaluation.

The new behavior is tested by the integration test for the
`ScriptCompiler` and also by the `Task` unit test which was rewritten
to use the `ScriptCompiler`.