(PUP-7894) Add tests for new V2 Scheduled Task provider
This commit changes the unit tests for the scheduled_task provider by running
the tests against both providers (win32_taskscheduler and
taskscheduler_api2). This ensures that any behaviors are consistent in both
providers.

The concrete class which services each provider is refactored into the
concrete_klass variable which is then used in the tests themselves.