(PUP-7946) Make ScriptCompiler take a block to execute after manifest
In order to first load and evaluate either a snippet or code or a
manifest the ScriptCompiler now accepts a block that takes an instance
of compiler as an argument. When present it is called with the script
compiler instance. It can then do things with its topscope, evaluate new
code etc.