(PUP-7946) Make the Puppet::Pal API use Ruby 2 keyword args
Some of the methods in the Pal API will need multiple variables and all
of them does not need to be specified. The readability and handling of
the old options Hash style is so much worse that it warrants creating a
new API that does not support 1.9.3.

If ever needed later, an uglier API "Pal_19" can be added.