(PUP-7845) Unmask certificate verify errors
If the TLS connection verification failed while verifying the peer
certificate chain, TLS verification related errors could be mistaken as
connection identity checking errors. For instance if an intermediate CA
certificate was received and was not trusted, verification would fail but
Puppet would report that the identity of the intermediate CA certificate
did not match the expected host.

For example the following error:

    Error: Could not request certificate: Server hostname 'sky' did not match server certificate; expected intermediateca.example.org

Was actually the following:

    "unable to get issuer certificate for /C=US/ST=Oregon/O=Puppet/CN=intermediateca.example.org/emailAddress=intermediateca@example.org"

This issue is resolved by checking for SSL verification errors first,
and performing hostname identity checking after the TLS certificates
have been verified.