(PUP-7946) Add StaticDirectory env loader and DRY up PAL API
This adds a StaticDirectory environment loader for use by the
PAL API when it is asked to treat a given directory as an environment
(irrespective of its directory name) and without forcing this env
directory to be in a directory on a environmentpath apointing only
directories that are environments.

This also DRYs up the PAL API wrt common routines used in both
`in_tmp_environment` and `in_environment` (the later using the
new StaticDirectory loader).

Unit tests are pending.