(PUP-7946) Add loading of existing env to PAL API
This makes it possible to use a given directory as an existing env with
a given name (that does not have to match the name of the directory).
If no modulepath is given a modulepath of ./modules relative to the
given directory will be used. This can be overridden by giving a
different (or empty `modulepath` argument to `in_environment`.
The resulting configured environment will load functions and plans in
this environment.