(PUP-7959) Fix apache2 dependent package list in service test
On SLES, the apache2-worker package is not a dependency of apache2, and
so doesn't get installed at the start of this test. Our new Power8
platform is based on SLES 12 SP2 and now causes zypper to return a
non-zero error when you explicitly try to uninstall a package that is
not already installed, which fails this test without this fix.