(PUP-7956) Allow declaration of hashes without surrounding curly braces
This commit changes the parser so that it will recognize hashes just by
detecting comma separated `<key> => <value>` bindings. A sequence of
such bindings is automatically converted into one single hash. This
means that the surrounding `{}` are no longer required in many places.

Sequences of bindings can be freely intermixed with other values in a
comma separated list. This is different from Ruby where such a sequence
is only allowed last. E.g. the following is legal

    notice('hello', a => 'word', b => 'universe', 'the sun is shining')

and will notice:

    "hello {a => word, b => universe} the sun is shining"

The detection of binding arguments is limited to literal lists, argument
lists, and arguments in access expressions. Binding arguments are not
allowed in a top level function call unless the argument list is
delimited by parenthesis.