(PUP-7875) Add ability to call run_task by name and type
This expands the set of signatures for run_type so that it can be called
with either a Type[Task], a Task, or a String task-name. The Type[Task]
and String variants take a node, or array (nested) of nodes and then
optionally a hash of named arguments to the task.