(PUP-7875) Simplify getting the arguments as hash from a Task
Before this, it was required to understand which type of Task being
operated on. A GenericTask has generic arguments in 'args', whereas a
specified Task has attributes available as a _pcore_init_hash.

This difference should not be significant when using the args.

This commit adds a `task_args` that gives both kind of tasks the same
API for getting the "task arguments".