(PUP-7308) Also propagate class tags to completed whit
If we do not propagate the class's tags to the completed whit, log
messages are still issued that the class's resources are being skipped
and their refresh events unscheduled, even though those events actually
occur. This commit sets the tags on the completed whit in addition to
the admissible whit, to make sure tags are propagated everywhere when
containers are replaced.