(PUP-7517) Add global timeout for agent runs
On occasion, a puppet agent can end up waiting indefinitely on some process
that will never return or terminate. This patch adds a new setting, runtimeout,
that can be used to specify the maximum duration allowed for a puppet run. When
set to a non-zero value, agent run that exceeds the given limit will be
interrupted and fail.