(docs) Align CONTRIBUTING.md with Travis CI.
When creating a commit summary, Travis CI can fail claiming it doesn't
match CONTRIBUTING.md guidelines:

```
This commit summary didn't match CONTRIBUTING.md guidelines:
  (DOCUMENT-729) Add context to static_catalogs setting.
The commit summary (i.e. the first line of the commit message) should
start with one of:
  (PUP-<digits>) # this is most common and should be a ticket at
tickets.puppet.com
  (docs)
  (docs)(DOCUMENT-<digits>)
  (maint)
  (packaging)
This test for the commit summary is case-insensitive.
```

Of these five rules for the start of a commit message, only the first,
`(PUP-<digits>)`, is documented in CONTRIBUTING.md, and is documented
ambiguously; it is unclear that only the PUP Jira project is accepted,
and that others will fail.

The second, `(docs)`, is listed in CONTRIBUTING incorrectly as `(doc)`.

The rest — `(docs)(DOCUMENT-<digits>)`, `(maint)`, and `(packaging)` —
are not documented at all.

This is a problem because a contributor might be unable to write a
valid, complete, and representative commit message despite following
the guidelines in CONTRIBUTING.md, and the message Travis CI produces
upon failure suggests rules that do not exist in CONTRIBUTING.md.

Resolve this issue by editing and adding to the rules in
CONTRIBUTING.md to match the output of the Travis CI error message.