(maint) Skip autoload tests on masters
The tests for autoloading resource types from generated types made the
assumption that Puppet masters are not Puppet agents. This assumption
depends on the beaker configuration used, and in many cases a master is
also classified as an agent. In this particular test the log files for
the agents must be empty while the log files for the master must be
populated, or vice versa.

Since the existing tests already have special cases for the master, this
commit resolves the issue by excluding the master from agent tests.