(PUP-7983) Enable testing system openssl linked PA on RHEL7 in CI pipes
Update openssl package to desired version on RHEL7 hosts since the default
version installed in images used by CI/VMPooler are not compatible with
agent setup and operation causing failures. Note that PA is built on centos7
which is also used to test on rhel platforms.
The openssl version to upgrade to can be specified using this environment
variable: RHEL7_OPENSSL_VERSION.
If not specified then it will fall back to a default openssl package version
that is known to work with PA linked against system openssl.