(PUP-8009) Only try to require gettext-setup once
A recent refactor of the gettext initialization code caused the require
for gettext-setup and locale to be called repeatedly during module
loading. This commit updates the gettext configuration module to only
require those gems once, and prevent unnecessary searching of the load
path.

Note, this reverts the wholesale gettext disable introduced previously.